WebAssembly: add start/end objects for metadata
The WebAssembly linker, unlike ELF linkers, doesn't define
__start and __stop symbols for sections. So we have to make our own.

If you put an object first in the command line, its symbols are placed first;
same for last in command line. So we make two files.

This is the same method that Swift used to use on ELF platforms.

run ./buildstartend.sh; these two objects must be linked at the start
and the end. see the updated linkPlease.sh file.