Add wasm branch scheme to update-checkout-config
This allows cloning all of the repositories directly with
./swift/utils/update-checkout --clone --scheme wasm
without relying on paths hardcoded in swiftwasm-sdk.
It also clones icu that way as well, as that one is used when building on Linux.
Later we could add "WebAssembly" platform to update-checkout-config.json to clone
it only for WebAssembly and Linux platforms, but for now it's pulled for all
platforms to make things easy.

https://github.com/swiftwasm/swift/pull/1